chore: moves module into its own directory
