+++ apps/webapp/src/script/application-periodic-checks/startApplicationPeriodicChecks.ts
